+++ src/Lens.ts
@@ -28,6 +28,7 @@ import { Option } from 'fp-ts/lib/Option'
 import { pipe } from 'fp-ts/lib/pipeable'
 import { Traversable1 } from 'fp-ts/lib/Traversable'
 import * as _ from './internal'
+import { Iso } from './Iso'
 import { Optional } from './Optional'
 import { Prism } from './Prism'
 import { Traversal } from './Traversal'
@@ -60,7 +61,7 @@ export const id: <S>() => Lens<S, S> = _.lensId
 // -------------------------------------------------------------------------------------
 
 /**
- * View a `Lens` as a `Optional`
+ * View a `Lens` as a `Optional`.
  *
  * @category converters
  * @since 2.3.0
@@ -68,7 +69,7 @@ export const id: <S>() => Lens<S, S> = _.lensId
 export const asOptional: <S, A>(sa: Lens<S, A>) => Optional<S, A> = _.lensAsOptional
 
 /**
- * View a `Lens` as a `Traversal`
+ * View a `Lens` as a `Traversal`.
  *
  * @category converters
  * @since 2.3.0
@@ -80,7 +81,7 @@ export const asTraversal: <S, A>(sa: Lens<S, A>) => Traversal<S, A> = _.lensAsTr
 // -------------------------------------------------------------------------------------
 
 /**
- * Compose a `Lens` with a `Lens`
+ * Compose a `Lens` with a `Lens`.
  *
  * @category compositions
  * @since 2.3.0
@@ -88,7 +89,17 @@ export const asTraversal: <S, A>(sa: Lens<S, A>) => Traversal<S, A> = _.lensAsTr
 export const compose: <A, B>(ab: Lens<A, B>) => <S>(sa: Lens<S, A>) => Lens<S, B> = _.lensComposeLens
 
 /**
- * Compose a `Lens` with a `Prism`
+ * Compose a `Lens` with a `Iso`.
+ *
+ * @category compositions
+ * @since 2.3.8
+ */
+export const composeIso: <A, B>(ab: Iso<A, B>) => <S>(sa: Lens<S, A>) => Lens<S, B> =
+  /*#__PURE__*/
+  flow(_.isoAsLens, compose)
+
+/**
+ * Compose a `Lens` with a `Prism`.
  *
  * @category compositions
  * @since 2.3.0
@@ -96,13 +107,22 @@ export const compose: <A, B>(ab: Lens<A, B>) => <S>(sa: Lens<S, A>) => Lens<S, B
 export const composePrism: <A, B>(ab: Prism<A, B>) => <S>(sa: Lens<S, A>) => Optional<S, B> = _.lensComposePrism
 
 /**
- * Compose a `Lens` with an `Optional`
+ * Compose a `Lens` with an `Optional`.
  *
  * @category compositions
  * @since 2.3.0
  */
-export const composeOptional = <A, B>(ab: Optional<A, B>) => <S>(sa: Lens<S, A>): Optional<S, B> =>
-  _.optionalComposeOptional(ab)(asOptional(sa))
+export const composeOptional = <A, B>(ab: Optional<A, B>): (<S>(sa: Lens<S, A>) => Optional<S, B>) =>
+  flow(asOptional, _.optionalComposeOptional(ab))
+
+/**
+ * Compose a `Lens` with an `Traversal`.
+ *
+ * @category compositions
+ * @since 2.3.8
+ */
+export const composeTraversal = <A, B>(ab: Traversal<A, B>): (<S>(sa: Lens<S, A>) => Traversal<S, B>) =>
+  flow(asTraversal, _.traversalComposeTraversal(ab))
 
 // -------------------------------------------------------------------------------------
 // combinators
@@ -137,7 +157,7 @@ export function modifyF<F>(F: Functor<F>): <A>(f: (a: A) => HKT<F, A>) => <S>(sa
 }
 
 /**
- * Return a `Optional` from a `Lens` focused on a nullable value
+ * Return a `Optional` from a `Lens` focused on a nullable value.
  *
  * @category combinators
  * @since 2.3.0
@@ -156,7 +176,7 @@ export function filter<A>(predicate: Predicate<A>): <S>(sa: Lens<S, A>) => Optio
 }
 
 /**
- * Return a `Lens` from a `Lens` and a prop
+ * Return a `Lens` from a `Lens` and a prop.
  *
  * @category combinators
  * @since 2.3.0
@@ -164,7 +184,7 @@ export function filter<A>(predicate: Predicate<A>): <S>(sa: Lens<S, A>) => Optio
 export const prop: <A, P extends keyof A>(prop: P) => <S>(sa: Lens<S, A>) => Lens<S, A[P]> = _.lensProp
 
 /**
- * Return a `Lens` from a `Lens` and a list of props
+ * Return a `Lens` from a `Lens` and a list of props.
  *
  * @category combinators
  * @since 2.3.0
@@ -174,7 +194,7 @@ export const props: <A, P extends keyof A>(
 ) => <S>(sa: Lens<S, A>) => Lens<S, { [K in P]: A[K] }> = _.lensProps
 
 /**
- * Return a `Lens` from a `Lens` and a component
+ * Return a `Lens` from a `Lens` and a component.
  *
  * @category combinators
  * @since 2.3.0
@@ -184,7 +204,7 @@ export const component: <A extends ReadonlyArray<unknown>, P extends keyof A>(
 ) => <S>(sa: Lens<S, A>) => Lens<S, A[P]> = _.lensComponent
 
 /**
- * Return a `Optional` from a `Lens` focused on a `ReadonlyArray`
+ * Return a `Optional` from a `Lens` focused on a `ReadonlyArray`.
  *
  * @category combinators
  * @since 2.3.0
@@ -193,7 +213,7 @@ export const index = (i: number) => <S, A>(sa: Lens<S, ReadonlyArray<A>>): Optio
   pipe(sa, asOptional, _.optionalComposeOptional(_.indexReadonlyArray<A>().index(i)))
 
 /**
- * Return a `Optional` from a `Lens` focused on a `ReadonlyRecord` and a key
+ * Return a `Optional` from a `Lens` focused on a `ReadonlyRecord` and a key.
  *
  * @category combinators
  * @since 2.3.0
@@ -202,7 +222,7 @@ export const key = (key: string) => <S, A>(sa: Lens<S, Readonly<Record<string, A
   pipe(sa, asOptional, _.optionalComposeOptional(_.indexReadonlyRecord<A>().index(key)))
 
 /**
- * Return a `Lens` from a `Lens` focused on a `ReadonlyRecord` and a required key
+ * Return a `Lens` from a `Lens` focused on a `ReadonlyRecord` and a required key.
  *
  * @category combinators
  * @since 2.3.0
@@ -211,7 +231,7 @@ export const atKey = (key: string) => <S, A>(sa: Lens<S, Readonly<Record<string,
   pipe(sa, compose(_.atReadonlyRecord<A>().at(key)))
 
 /**
- * Return a `Optional` from a `Lens` focused on the `Some` of a `Option` type
+ * Return a `Optional` from a `Lens` focused on the `Some` of a `Option` type.
  *
  * @category combinators
  * @since 2.3.0
@@ -221,7 +241,7 @@ export const some: <S, A>(soa: Lens<S, Option<A>>) => Optional<S, A> =
   composePrism(_.prismSome())
 
 /**
- * Return a `Optional` from a `Lens` focused on the `Right` of a `Either` type
+ * Return a `Optional` from a `Lens` focused on the `Right` of a `Either` type.
  *
  * @category combinators
  * @since 2.3.0
@@ -231,7 +251,7 @@ export const right: <S, E, A>(sea: Lens<S, Either<E, A>>) => Optional<S, A> =
   composePrism(_.prismRight())
 
 /**
- * Return a `Optional` from a `Lens` focused on the `Left` of a `Either` type
+ * Return a `Optional` from a `Lens` focused on the `Left` of a `Either` type.
  *
  * @category combinators
  * @since 2.3.0
@@ -241,7 +261,7 @@ export const left: <S, E, A>(sea: Lens<S, Either<E, A>>) => Optional<S, E> =
   composePrism(_.prismLeft())
 
 /**
- * Return a `Traversal` from a `Lens` focused on a `Traversable`
+ * Return a `Traversal` from a `Lens` focused on a `Traversable`.
  *
  * @category combinators
  * @since 2.3.0